mlrecogniser/requirements.txt
grpcio
grpcio-tools
torch
numpy
scipy 